+++ office365/migration/sharepoint/scanners/__init__.py
@@ -5,9 +5,18 @@ from office365.migration.assessment.scanners.base import (
     BaseScanner,
     ScanTarget,
 )
+from office365.migration.sharepoint.scanners.browser_file_handling import (
+    BrowserFileHandlingRecord,
+    BrowserFileHandlingScanner,
+)
+from office365.migration.sharepoint.scanners.checked_out_files import (
+    CheckedOutFilesRecord,
+    CheckedOutFilesScanner,
+)
 from office365.migration.sharepoint.scanners.fields import FieldScanner
 from office365.migration.sharepoint.scanners.file_versions import FileVersionsRecord, FileVersionsScanner
 from office365.migration.sharepoint.scanners.files import FileScanner
+from office365.migration.sharepoint.scanners.large_excel_files import LargeExcelFilesRecord, LargeExcelFilesScanner
 from office365.migration.sharepoint.scanners.large_sites import LargeSitesScanner
 from office365.migration.sharepoint.scanners.locked_sites import SiteLockedScanner
 from office365.migration.sharepoint.scanners.paths import PathScanner
@@ -17,10 +26,16 @@ from office365.migration.sharepoint.scanners.summary import SiteScanSummary
 __all__ = [
     "AssessmentOptions",
     "BaseScanner",
+    "BrowserFileHandlingRecord",
+    "BrowserFileHandlingScanner",
+    "CheckedOutFilesRecord",
+    "CheckedOutFilesScanner",
     "FieldScanner",
     "FileScanner",
     "FileVersionsRecord",
     "FileVersionsScanner",
+    "LargeExcelFilesRecord",
+    "LargeExcelFilesScanner",
     "LargeSitesScanner",
     "PathScanner",
     "PermissionScanner",